make audioview look nice
layout is now simpler using a fixed size analyzer below tags. in addition
when switching to the audioview the window resizes to the minimum size
(maximum size of analyzer + chrome). this now gives a very nice
appearance and if the user chooses to resize the window the entire view is
centered.

additionally:
- explicitly set fontsize via kglobalsettings
- stop adding the track number to the track label (looks silly)