Call KLocalizedString::setApplicationDomain("dragonplayer");
dragonplayer is using
   -DTRANSLATION_DOMAIN=\"dragonplayer\"
that is wrong for applications, apps need to use
   KLocalizedString::setApplicationDomain
since that is needed for some things like the transltators tab in the about
dialog.

I have not removed -DTRANSLATION_DOMAIN=\"dragonplayer\" because on the other hand
this code is also a library (i.e. the dragonplayer kpart).

The "being pedanticly correct" fix would be using -DTRANSLATION_DOMAIN=\"dragonplayer\"
only to compile the files that are part of the kpart but that's prone to
break at some point and using both KLocalizedString::setApplicationDomain and
-DTRANSLATION_DOMAIN does not create any problem so i went that route

CCMAIL: lukas@kde.org