Fix: user settings for gui not saved after program is closed
Summary:
When the player switches to full screen mode the settings for the visibility of menubar, toolbar and statusbar are set to fixed values and the user settings are saved in backup variables.
These backup values are restored when the player is closed. If the player is not in full screen mode, the backup variable have default values and the user settings are overwritten.

Solution: Restore settings only when the window is in fullscreen mode.

BUG: 376961

Reviewers: sitter

Reviewed By: sitter

Differential Revision: https://phabricator.kde.org/D15594