Add ability to copy the URL of played media
Summary:
FEATURE: 409743

SUMMARY

STEPS TO REPRODUCE
1. Click "Play Stream" or "Play File"
2. Enter stream URL / select file
3. Once the stream / file starts playing, try to extract the URL of the playing stream / file.

OBSERVED RESULT
If the user forgets the location of the file or URL of the stream, there  is no way to get the URL once the media is played.

EXPECTED RESULT
The context menu of the recently played items should provide option to copy the stream URL / file location. Will be handy if the user forgets the original stream URL / file location.

Test Plan: No test plan as this is a minor change.

Reviewers: kossebau, cfeck, lukas, sitter

Reviewed By: sitter

Differential Revision: https://phabricator.kde.org/D22423