Fixup bf1ece77
That caused popups to close when the user was still interacting with
them in full screen mode. Revert that.

Now any popups being open is simply another condition for remaining in
the "fullscreen-active" state. Once all popups have been closed by the
user, moving into the "fullscreen" state is unblocked.