port to qtquick
- the UI is now written using qtquick. it's largely a replication of the
widget UI with some simplifications to closer resemble the simplicity of
the original codeine player design. offering a hopefully more
streamlined UX we can iterate on
- now using qtmultimedia instead of phonon. with the advance of the
ffmpeg backend qtmultimedia should work more reliably and we decided to
deprecate phonon anyway. while I had a libvlc based prototype for a
while it ultimately didn't amount to anything because the release
cadence of libvlc is too slow for us
- mpris support has been adjusted accordingly
- audio player view has been removed. it always felt awkward and I
suspect it wasn't really used
- as part of this the analyzer code base is no more
- dvd support has been removed. I have no hardware and it was broken for
a while already without anyone fixing it
- load view has been removed to simplify the UI and UX. seeing as
without disc support we only ever open urls the view wasn't too useful
anymore but had a lot of code and special handling. this effectively
means that the new page stack only needs to deal with one page (well,
and the about page, but that doesn't really count since that always sits
on top)

Co-authored-by: Nate Graham <nate@kde.org>