stop using multimediaprivate
it's a bit of an unnecessary dep since all we did was access a QObject
base. we can just as well reinterpret from a forward declared pointer to
QObject* instead. while that is a bit meh because of the cast it saves
us the entire private code so it's probably better overall

(the private usage was a bit shoddy anyway because the link_lib and
include were not conditional)